requirements.txt
pyside6
pyyaml
qt-material
websocket-client